requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.2
Flask-SQLAlchemy==2.0
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.9
Tempita==0.5.2
Werkzeug==0.10.1
aniso8601==0.92
decorator==3.4.2
itsdangerous==0.24
pbr==0.10.8
pytz==2015.2
six==1.9.0
sqlalchemy-migrate==0.9.5
sqlparse==0.1.14
